requirements.txt
Flask
pandas
scikit-learn
numpy
requests
schedule
python-dotenv
cohere
python-docx
PyPDF2